cmdline: split serializer module into separate file
